refactor inteview-report
