+++ web/src/features/interview-report/shared/types.ts
@@ -1,0 +1,6 @@
+import type { Database } from "@mirai-gikai/supabase";
+
+export type InterviewReport =
+  Database["public"]["Tables"]["interview_report"]["Row"];
+export type InterviewReportInsert =
+  Database["public"]["Tables"]["interview_report"]["Insert"];
